requirements.txt
streamlit
pillow
transformers
torch
torchvision
numpy
pandas
scikit-learn
tqdm
datasets
matplotlib